Making Things more organized Test are ok to all platforms missing github test Plan
